refs #21630 rm unneded comments from James' MaxEnt code to Mantid
