+++ Framework/PythonInterface/plugins/algorithms/Muon/MuonMaxentTools/opus.py
@@ -1,0 +1,14 @@
+import numpy as np
+
+
+def OPUS(x,SAVETIME_i2,PULSESHAPE_convol,DETECT_a,DETECT_b,DETECT_e):
+	ngroups=DETECT_a.shape[0]
+	npts=DETECT_e.shape[0]
+	n=x.shape[0]
+	y=np.zeros([SAVETIME_i2],dtype=np.complex_)
+	y[:n]=x*PULSESHAPE_convol
+	y2=np.fft.ifft(y)*SAVETIME_i2 # SN=+1, inverse FFT without the 1/N
+	ox=(np.outer(np.real(y2[:npts]),DETECT_a)+np.outer(np.imag(y2[:npts]),DETECT_b))*DETECT_e[:,np.newaxis]
+	return ox
+	
+
